# lib/api-zod/src/generated/types/quizInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Study Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { QuizInputDifficulty } from './quizInputDifficulty';

export interface QuizInput {
  /** @minLength 1 */
  notes: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  numQuestions?: number;
  difficulty?: QuizInputDifficulty;
}