# lib/api-zod/src/generated/types/todo.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Study Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { TodoPriority } from './todoPriority';

export interface Todo {
  id: number;
  title: string;
  completed: boolean;
  priority: TodoPriority;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}