# lib/api-zod/src/generated/types/todoStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Study Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface TodoStats {
  total: number;
  completed: number;
  pending: number;
  highPriority: number;
}